backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
geopandas==0.14.3
shapely==2.0.3
pandas==2.2.1
numpy==1.26.4
xgboost==2.0.3
scikit-learn==1.4.1
joblib==1.3.2
osmnx==1.9.1
networkx==3.3
requests==2.31.0
pyarrow==16.0.0
python-dotenv==1.0.1
